Use get_table_from_query to find table names
There's a bit of juggling needed with placeholders because we're dealing with un-interpolated PHP strings containing variable names, rather than  the interpolated literal result.

This basically assumes that any variable used as a table name in a query must inherently be safe-ish, and therefore adds it to the list of `warn_only_parameters`.