Use a constant table name in this test
Turns out this is trickier than I initially thought.

The original test query isn't really safe because it uses `$table` for the table name. That ought to qualify it as a warning rather than an error (which means it wasn't suitable in the safe tests as written). But in order to work that out, `is_warning_parameter()` and `unwind_unsafe_assignments()` would need to be substantially rewritten or replaced with something that does a bit of both.

Since that's pretty complex, I'll just make the query safe and fix the other issues with duplicate context messages for now.