Add a unit test for a tricky false positive bug
This is hard to fix because `is_warning_parameter()` doesn't recurse backwards. It's only run on `$query` (line 446) and not the previous line where it would correctly know that `{$table_name}` should only trigger a warning.

Fixing this probably needs an overhaul of the `mark_*()` functions. That would have other benefits too.